Restrict branch builds from forks, hash actions
Updates the container branch publishing job to only allow publishing
containers from branches in this repository. This ensures PRs from forks
aren't allow to publish to this repositories container registry.

Bumps the checkout, setup-go, and setup-ko actions to the latest
version, and ensures all ko-build-* jobs are using the hash.

Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>