fix commenting on ruleset.yaml
the yaml comments were being inlined as part of the string
on the line above, not a proper comment. The else clause has
been fixed to prevent this.

However, the comment *also* gets thrown away, so change it over
to a helm template comment instead since that is the only place
it will be seen.

Signed-off-by: Jordan Evans <jevans@linuxfoundation.org>